% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitTests-methods.R
\name{dt-methods}
\alias{dt-methods}
\alias{print.dt.distr}
\alias{print.dt.rv}
\alias{plot.dt.distr}
\alias{plot.dt.rv}
\title{Methods for da objects}
\usage{
\method{print}{dt.distr}(x, digits = getOption("digits"), ...)

\method{print}{dt.rv}(x, digits = getOption("digits"), ...)

\method{plot}{dt.distr}(x, ...)

\method{plot}{dt.rv}(x, ...)
}
\arguments{
\item{x}{an object of class \code{da} as returned by one of the package functions.}

\item{digits}{the number of digits to round to.}

\item{...}{further arguments, currently ignored.}
}
\value{
The \code{print} methods simply print and return nothing.
}
\description{
Methods defined for objects returned from the \code{\link{distr.test}}, \code{\link{distr.btest}}, and \code{\link{rv.test}} functions.
}
